+++ yash-semantics/CHANGELOG.md
@@ -7,12 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.9.0] - Unreleased
 
+### Added
+
+- Support for the `pipefail` shell option in pipeline execution
+    - When this option is enabled, the exit status of a pipeline reflects the
+      failure of any command in the pipeline, not just the last command.
+    - This is implemented in
+      `impl command::Command for yash_syntax::syntax::Pipeline`.
+
 ### Changed
 
 - The special parameter `!` is now considered unset if no asynchronous command
   has been executed, that is, if `JobList::last_async_pid()` is zero.
     - This change is observed in the results of the `expand` method of
       `impl expansion::initial::Expand for yash_syntax::syntax::TextUnit`.
+- External dependency versions:
+    - yash-env 0.8.0 → 0.8.1
 
 ## [0.8.1] - 2025-09-20